backend/requirements.txt
chainlit
semantic-kernel>=1.27.2
openai
python-dotenv
pydantic
requests
